pyproject.toml
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
[build-system]
requires = ["pip>=19.1.1", "setuptools", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "fmu-dataio"
description = "Facilitate data io in FMU with rich metadata"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache 2.0" }
authors = [{ name = "Equinor", email = "xxxxx@equinor.com" }]
keywords = ["fmu", "sumo"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dynamic = ["version"]
dependencies = [
    "fmu-config>=1.1.0",
    "fmu-datamodels~=0.14.0",
    "numpy",
    "pandas",
    "pyarrow",
    "pydantic",
    "PyYAML",
    "xtgeo>=2.16",
    "fmu-sumo",
    "jsonschema",
]

[project.urls]
Homepage = "https://github.com/equinor/fmu-dataio"
Repository = "https://github.com/equinor/fmu-dataio"
Issues = "https://github.com/equinor/fmu-dataio/issues"
Documentation = "https://fmu-dataio.readthedocs.io"

[project.optional-dependencies]
dev = [
    "coverage>=4.1",
    "ert",
    "hypothesis",
    "jsonschema",
    "mypy",
    "pandas-stubs",
    "pyarrow-stubs",
    "pydocstyle",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-runner",
    "pytest-xdist",
    "rstcheck",
    "ruff",
    "types-PyYAML",
    "types-requests",
]
docs = [
    "Sphinx",
    "autoapi",
    "autodoc-pydantic>=2.0.0",
    "furo",
    "myst-parser",
    "pydocstyle",
    "sphinx-autodoc-typehints",
    "sphinx-copybutton",
    "sphinx-togglebutton",
    "sphinxcontrib-apidoc",
    "urllib3<1.27",
]

[project.entry-points.ert]
dataio_case_metadata = "fmu.dataio.scripts.create_case_metadata"
dataio_copy_preprocessed = "fmu.dataio.scripts.copy_preprocessed"

[tool.setuptools_scm]
write_to = "src/fmu/dataio/version.py"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = "tests"

[tool.coverage.run]
omit = [
    "version.py",
    "tests/**",
]

[tool.pydocstyle]
convention = "google"
match = '(?!(test_|_)).*\.py'

[tool.ruff]
line-length = 88
exclude = ["version.py", "docs/ext/**"]

[tool.ruff.lint]
ignore = [
    "B028",  # No explicit `stacklevel` keyword argument found
    "C901",  # mccabe complex-structure
    "PD011",  # Avoid using .values on dataframes (gives false positives)
    "TC001",  # Move imports under TYPE_CHECKING. Breaks Pydantic annotations
    "TC003",  # Move stdlib imports under TYPE_CHECKING. Breaks Pydantic annotations
]
select = [
    "B",  # flake-8-bugbear
    "C",  # pylint-convention
    "C90",  # mccabe
    "E",  # pycodestyle-error
    "F",  # pyflakes
    "I",  # isort
    "NPY",  # numpy
    "PD",  # pandas-vet
    "PIE",  # flake8-pie
    # "PL",  # pylint
    "Q",  # flake8-quotes
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    # "TID",  # flake8-tidy-imports
    "UP",  # pyupgrade
    "W",  # pylint-warnings
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
disallow_untyped_defs = true
exclude = "^((tests|docs|examples|build|tools)/|conftest.py?)"
extra_checks = true
ignore_missing_imports = true
python_version = 3.11
strict_equality = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
